% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{alignSets}
\alias{alignSets}
\title{Aligng datasets with procrustes analyses}
\usage{
alignSets(set1, set2, dilation = TRUE, translation = TRUE)
}
\arguments{
\item{set1}{M x D [matrix] with M samples and D dimensionality reduced
data P; the matrix to be transformed.}

\item{set2}{M x D [matrix] with M samples and D dimensionality reduced
data P (with the same sample order as set1); the target matrix.}

\item{dilation}{[logical] indicating whether set1 should be dilated.}

\item{translation}{[logical] indicating whether set1 should be translated.}
}
\value{
[matrix] that is the Procrustes transformed version of set1.
}
\description{
Aligng datasets with procrustes analyses
}
