% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{analyseCorrelation}
\alias{analyseCorrelation}
\title{Order results based on highest correlation}
\usage{
analyseCorrelation(mat, verbose = FALSE)
}
\arguments{
\item{mat}{[D x D] matrix with correlation coefficients from column-wise 
comparison of D lower dimensions}

\item{verbose}{[logical] If set, progress messages are printed to standard 
out}
}
\value{
named list with ordering of the correlations (reorder) and the 
maximum correlations (maxcor)
}
\description{
Order results based on highest correlation
}
