% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{computeDimReduction}
\alias{computeDimReduction}
\title{Compute dimensionality reduction}
\usage{
computeDimReduction(method, Y, optN = NULL, ndim = NULL, kmin = 1,
  kmax = 40, verbose = FALSE)
}
\arguments{
\item{method}{Dimensionality reduction method [string] to be applied; one of
DiffusionsMaps, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA, 
nMDS, tSNE and PEER}

\item{Y}{[N x P] data matrix for which the dimensionality of P should be
reduced}

\item{optN}{optimal number [int] of neighbours to consider for dimensionality
reduction; relevant for methods LLE, LaplacianEigenmaps, Isomap and tSNE. If 
not provided, will be estimated via \code{\link{lle::calc_k}}.}

\item{ndim}{maximum dimensionality [int] to retain in the data; large values
can cause long computation times; if not provided max(P,N) is chosen.}

\item{kmin}{if optN is not provided, kmin [int] specifies the minimum number
of neighbours supplied to  \code{\link{lle::calc_k}}}

\item{kmax}{if optN is not provided, kmax [int] specifies the maximum number
of neighbours supplied to  \code{\link{lle::calc_k}}}

\item{verbose}{[logical] If set, progress messages are printed to standard 
out}
}
\value{
named list of results from dimensionality reduction; if multiple 
methods were specified, each named list entry contains the following:
Y_red:  named list with dimensionality reduced phenotypes (reducedY) and 
object returned by specified dimensionality reduction method (results) with
additional output
M: vector [double] with Trustworthiness and Continuity estimates for the
dimensionality reduction
}
\description{
Compute dimensionality reduction
}
