% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{correlateSets}
\alias{correlateSets}
\title{Column-wise correlation of two datasets}
\usage{
correlateSets(set1, set2, type = "spearman")
}
\arguments{
\item{set1}{[M x D] matrix with M samples and D dimensionality reduced
data P}

\item{set2}{[M x D] matrix with M samples and D dimensionality reduced
data P, with the same sample order as set1}

\item{type}{type [string] of correlation to use, one of spearman or pearson}
}
\value{
named list with correlation coefficient (r) and p-value of 
correlation (p) for each column-column correlation
}
\description{
Column-wise correlation of two datasets
}
