% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{formatComparison}
\alias{formatComparison}
\title{Format the comparison results}
\usage{
formatComparison(comparison_list)
}
\arguments{
\item{comparison_list}{output [list] from resulsComparison containing with 
the size of the overlapping samples between samples sets (size), the
reordered correlation coefficients (reorder), the maximum correlation
(maxcor), the original column-wise correlation coefficients 
(componentxcomponent_r) and the reordered column-wise correlation
coefficients (componentxcomponent_r_reordered)}
}
\value{
list with same entries as input list, formated to only contain unique
comparisons: size of the overlapping samples between samples 
sets (size), the reordered correlation coefficients (reorder), the maximum 
correlation (maxcor), the original column-wise correlation coefficients 
(componentxcomponent_r) and the reordered column-wise correlation
coefficients (componentxcomponent_r_reordered)
}
\description{
Extract relevant data and remove null-entries
}
