% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{formatStability}
\alias{formatStability}
\title{Format stability}
\usage{
formatStability(es, threshold)
}
\arguments{
\item{es}{[list]}

\item{threshold}{[double]}
}
\value{
dataframe with stability and component variables
}
\description{
Takes the output of \code{\link{estimateStability}} and returns stability
estimates per component for the specified threshold
}
