% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{invnorm}
\alias{invnorm}
\title{Inverse normalise}
\usage{
invnorm(x)
}
\arguments{
\item{x}{vector with [n] elements to normalise.}
}
\value{
x vector with [n] normalised elements.
}
\description{
Apply rank-based quantile normalistation to data.
}
\examples{
x <- seq(1,20,2)
x_norm <- invnorm(x)
}
