% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{medianCorr}
\alias{medianCorr}
\title{Compute median correlation of low-dimensional components across subsamples}
\usage{
medianCorr(es, threshold)
}
\arguments{
\item{es}{output [list] of \code{\link{estimateStability}} with named entries
[1] stability: list with one dataframe per threshold,
where each dataframe contains the stability estimate of the low-dimensional
components for that threshold value, [2] corr: data.frame with correlation of
the low-dimensional components for all pairwise comparisons and components.}

\item{threshold}{threshold [double] to mark median correlation as
above/below.}
}
\value{
Dataframe with median correlation for each component.
}
\description{
Takes the output of \code{\link{estimateStability}} and returns the median
correlation per component for the specified threshold.
}
