% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{methodsDimReduction}
\alias{methodsDimReduction}
\title{Wrapper function for dimensionality reduction methods}
\usage{
methodsDimReduction(Y, ndim, distY = dist(Y, method = dist.method),
  dist.method = "euclidean", method = c("DiffusionMap", "DRR", "ICA",
  "LLE", "Isomap", "LaplacianEigenmap", "MDS", "PCA", "kPCA", "nMDS",
  "tSNE", "UMAP"), optN = NULL, verbose = FALSE, params = NULL)
}
\arguments{
\item{Y}{\[N x P\] data matrix for which the dimensionality of P should be
reduced}

\item{ndim}{maximum dimensionality [integer] to retain in the data; large
values can cause long computation times.}

\item{distY}{[dist] object of class dist containing pairwise distances of Y
used for methods DiffusionMap, Isomap, MDS and nMDS; if non specified,
stats::dist with Euclidean distance applied to supplied Y.}

\item{dist.method}{[character] method for computing the distance matrix; one of
euclidean, maximum, manhattan, canberra, binary or minkowski; see
\code{\link[stats]{dist}} for details.}

\item{method}{Dimensionality reduction method [character] to be applied; one
of DiffusionsMaps, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA,
nMDS and tSNE.}

\item{optN}{optimal number [integer] of neighbours to consider for
dimensionality reduction; relevant for methods LLE, LaplacianEigenmaps,
Isomap and tSNE.}

\item{verbose}{[logical] If set, progress messages are printed to standard
out.}

\item{params}{[list] optional additional parameters for dimensionality
reduction methods; see details.}
}
\value{
named list with dimensionality reduced phenotypes (reducedY) and
object returned by specified dimensionality reduction method (results) with
additional output, see details.
}
\description{
Wrapper function for dimensionality reduction methods
}
\details{
methodsDimReduction wraps around the following implementations of
the dimensionality reduction methods it provides:
* Diffusion Map: \code{\link[diffusionMap]{diffuse}}
* Dimensionality reduction by regression (DRR): \code{\link[DRR]{drr}}
* Independent component analysis (ICA): \code{\link[fastICA]{fastICA}}
* Local liner embedding (LLE): \code{\link[lle]{lle}}
* Isomap: \code{\link[vegan]{isomap}}
* Laplacian Eigenmap: \code{\link[loe]{spec.emb}} and
  \code{\link[loe]{make.kNNG}}
* Multi-dimensional scaling (MDS) : \code{\link[stats]{cmdscale}}
* Principal component analysis (PCA): \code{\link[stats]{prcomp}}
* Kernel PCA (kPCA): \code{\link[kernlab]{kpca}}
* non-metrix MDS (nMDS): \code{\link[vegan]{metaMDS}}
* t- stochastic neighbourhood embedding (tSNE): \code{\link[Rtsne]{Rtsne}}
* Uniform manifold approximation and projection (umap):
   \code{\link[umap]{umap}}
}
