% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandlineFunctions.R
\name{prepInput}
\alias{prepInput}
\title{Prepare input files}
\usage{
prepInput(args)
}
\arguments{
\item{args}{[list] of command line arguments}
}
\value{
formated high-dimensional dataset matrix
}
\description{
Prepare input files for runEstimateStability, createSubsets and
runDimReduction. Check if files exist, read, format and -if provided-regress
covariates from highdim file.
}
