% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{resultsComparison}
\alias{resultsComparison}
\title{Apply compareSetup and format results}
\usage{
resultsComparison(data_list, procrustes, verbose = FALSE)
}
\arguments{
\item{data_list}{[list] output of \code{\link{subsetDimReduction}}. Each list
entry is a matrix with row and column names; row.names are crucial to allow
for the comparison of the lower dimensional embedding across common sample
sets.}

\item{procrustes}{[logical] indication whether a procrustes transformation
to match the lower dimensional representations of common samples between
subsets should be performed.}

\item{verbose}{[logical] If set, progress messages are printed to standard
out.}
}
\value{
named list with the size of the overlapping samples between samples
sets (size), the reordered correlation coefficients (reorder), the maximum
correlation (maxcor), the original column-wise correlation coefficients
(componentxcomponent_r) and the reordered column-wise correlation
coefficients (componentxcomponent_r_reordered)
}
\description{
Apply compareSetup and format results
}
