% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rmnulls}
\alias{rmnulls}
\title{Remove nulls from nested list}
\usage{
rmnulls(x)
}
\arguments{
\item{x}{[list] with [n] elements (NULL and non-NULL elements)}
}
\value{
[list] with [n - \# NULL elements] elements
}
\description{
Remove NULL elements from list.
}
\examples{
y <- list(a=1,b=2,c=NULL)
y_nnull <- rmnulls(y)
}
