% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandlineFunctions.R
\name{runCreateSubsets}
\alias{runCreateSubsets}
\title{Command line execution for Dimensionality reduction}
\usage{
runCreateSubsets()
}
\description{
runCreateSubsets runs without arguments. Upon call, it reads
command-line parameters pointing to the [highdim] file for which sample
subsets should be created. [nrCV] subsets of the [highdim] data of size
[sizeCV] are generated by random sampling of the row indeces (sampling seed
can be specified via [seed]). The subsets will be saved in the same directory
as the [highdim]_cv1.csv, ..., [highdim]_cv[nrCV].csv. For help on
the command line arguments that can be passed, see examples below. From the
command line, the help function can be called via
Rscript -e "Stability::runCreateSubsets()" --args --help
}
\examples{
# (not run)
# Rscript -e "Stability::runDimensionalityReduction()" \\
#--args \\
#--highdim=/path/to/highdimfile \\
#--cov=/path/to/optional/covariatefile \\
#--seed=234 \\
#--nrCV=10 \\
#--sizeCV=0.8 \\
#--showProgress \\
}
