% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{saveDimReduction}
\alias{saveDimReduction}
\title{Save results of dimensionality reduction}
\usage{
saveDimReduction(results, outdir)
}
\arguments{
\item{results}{named list of results from dimensionality reduction; if 
multiple methods were specified, each named list entry containing:
Y_red: named list with dimensionality reduced phenotypes (reducedY) and 
object returned by specified dimensionality reduction method (results) with
additional output.
M: vector [double] with Trustworthiness and Continuity estimates for the
dimensionality reduction.}

\item{outdir}{path/to/directory [string] where results will be saved.}
}
\description{
Save results of dimensionality reduction
}
