% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ss}
\alias{ss}
\title{Sum of squares}
\usage{
ss(data)
}
\arguments{
\item{data}{matrix of vector [float] to compute sum of squares for.}
}
\value{
value [float] of sum of squares.
}
\description{
Compute sum of squares of vector or matrix
}
\examples{
x <- 1:10
x_ss <- ss(x)
}
