% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandlineFunctions.R
\name{wrapDR}
\alias{wrapDR}
\title{Wrap dimensionality reduction functions}
\usage{
wrapDR(Y, args)
}
\arguments{
\item{Y}{[matrix] of high-dimensional dataset}

\item{args}{[list] of command line arguments
€return list with dimensionality reduction results}
}
\description{
Provide wrapper for dimensionality reduction functions in
runEstimateStability and runDimReduction. Check which analysis mode, run
dimensionality reduction and save output files.#' covariates from highdim file.
}
