% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{checkParams}
\alias{checkParams}
\title{Check optional dimensionaloty reduction method parameters}
\usage{
checkParams(params, method)
}
\arguments{
\item{params}{[list] of parameters to check; list names must be identical
to argument names of chosen dimensionality reduction method.}

\item{method}{Dimensionality reduction method [string] to be applied; one of
DiffusionMap, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA,
nMDS, tSNE, UMAP and PEER}
}
\value{
named list with default and specified parameters taken by specified
dimensionality reduction method
}
\description{
Checks if optional parameters passed to dimensionality reduction method are
valid and returns a list of all default and optinally set parameters.
}
\examples{
params_DRR <- checkParams(list(cv.folds=10), method="DRR")
params_ICA <- checkParams(list(fun="logcosh", maxit=500), method="ICA")
}
