% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{estimateStability}
\alias{estimateStability}
\title{Estimate stability}
\usage{
estimateStability(dr, threshold, verbose = FALSE)
}
\arguments{
\item{dr}{[list]}

\item{threshold}{[double]}

\item{verbose}{[logical] If set, progress messages are printed to standard 
out}
}
\value{
named list
}
\description{
Estimate stability
}
