% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{methodsDimReduction}
\alias{methodsDimReduction}
\title{Wrapper function for dimensionality reduction methods}
\usage{
methodsDimReduction(Y, ndim, method = c("DiffusionsMaps", "DRR", "ICA", "LLE",
  "Isomap", "LaplacianEigenmap", "MDS", "PCA", "kPCA", "nMDS", "tSNE", "PEER"),
  optN = NULL)
}
\arguments{
\item{Y}{[N x P] data matrix for which the dimensionality of P should be
reduced}

\item{ndim}{maximum dimensionality [int] to retain in the data; large values
can cause long computation times}

\item{method}{Dimensionality reduction method [string] to be applied; one of
DiffusionsMaps, DRR, ICA, LLE, Isomap, LaplacianEigenmap, MDS, PCA, kPCA, 
nMDS, tSNE and PEER}

\item{optN}{optimal number [int] of neighbours to consider for dimensionality
reduction; relevant for methods LLE, LaplacianEigenmaps, Isomap and tSNE.}
}
\value{
named list with dimensionality reduced phenotypes (reducedY) and 
object returned by specified dimensionality reduction method (results) with
additional output, see details
}
\description{
Wrapper function for dimensionality reduction methods
}
