% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stability.R
\name{resultsComparison}
\alias{resultsComparison}
\title{Apply compareSetup and format results}
\usage{
resultsComparison(data_list, verbose = FALSE)
}
\arguments{
\item{data_list}{[list]}

\item{verbose}{[logical] If set, progress messages are printed to standard 
out}
}
\value{
named list with the size of the overlapping samples between samples 
sets (size), the reordered correlation coefficients (reorder), the maximum 
correlation (maxcor), the original column-wise correlation coefficients 
(componentxcomponent_r) and the reordered column-wise correlation
coefficients (componentxcomponent_r_reordered)
}
\description{
Apply compareSetup and format results
}
