% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandlineFunctions.R
\name{runEstimateStability}
\alias{runEstimateStability}
\title{Command line execution for Stability}
\usage{
runEstimateStability()
}
\description{
runEstimateStability runs without arguments. Upon call, it reads command-line
parameters and supplies these to \code{\link{estimateStability}}.
It reads the dimensionality reduction results from the cross-validation (
Rscript -e "Stability::runDimReduction()") and analyses the stability of
the low-dimensional embedding via \code{\link{estimateStability}}. For
details on \code{\link{estimateStability}} please refer to the help pages.
For help on the command line arguments that can be passed, see examples
below. From the command line, the help function can be called via
Rscript -e "Stability::runEstimateStability()" --args --help
}
\examples{
# (not run)
# Rscript -e "Stability::runEstimateStability()" \\
#--args \\
#--highdim=/path/to/highdimfile \\
#--cov=/path/to/optional/covariatefile \\
#--seed=234 \\
#--nrCV=10 \\
#--neighbours=8 \\
#--method=DRR \\
#--name=Dimreduction_highdimfile \\
#--outdir=/tmp \\
#--showProgress \\
}
