% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{saveDimReduction}
\alias{saveDimReduction}
\title{Save results of dimensionality reduction}
\usage{
saveDimReduction(results, method, outdir = NULL, name = NULL,
  prefix = NULL)
}
\arguments{
\item{results}{named list of results from dimensionality reduction;
Y_red: named list with dimensionality reduced phenotypes (reducedY) and
object returned by specified dimensionality reduction method (results) with
additional output.
M: vector [double] with Trustworthiness and Continuity estimates for the
dimensionality reduction.}

\item{method}{[string] name of dimensionality reduction method; will be
appended to output file name.}

\item{outdir}{[string] full path/to/directory [string] where results will be
saved. Full file name will be created by outdir/name_method.csv.
Alternatively, prefix can be provided to create file prefix_method.csv.}

\item{name}{[string] name of output file}

\item{prefix}{[string] of output file name, containing full
path/to/directory [string] where results will be saved. Results will be saved
as prefix_method.csv}
}
\description{
Save results of dimensionality reduction
}
